refactor(step4): car 패키지 추가
