+++ src/main/java/step3/game/car/Car.java
@@ -1,4 +1,4 @@
-package step3.game;
+package step3.game.car;
 
 public class Car {
