+++ src/main/java/step3/game/car/CarName.java
@@ -1,4 +1,4 @@
-package step3.game;
+package step3.game.car;
 
 import java.util.Objects;
