+++ src/test/java/step3/game/CarTest.java
@@ -2,6 +2,8 @@ package step3.game;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.game.car.Car;
+import step3.game.car.CarName;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
