+++ src/test/java/step3/game/RacingCarGameTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.game.car.CarName;
 
 import java.util.HashSet;
 import java.util.Set;
